Add jest tests
