+++ src/package.json
@@ -3,9 +3,9 @@
   "version": "0.0.1",
   "description": "Brave Variations",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 0",
+    "test": "jest",
     "lint": "eslint . --ext ts,tsx",
-    "tracker": "node finch_tracker/build/finch_tracker/finch_tracker_main.js",
+    "tracker": "node finch_tracker/build/finch_tracker/main.js",
     "clean": "rm -rf ./proto/generated/proto_bundle* web/public/bundle/ finch_tracker/build/",
     "build:proto": "pbjs -t static-module --keep-case ./proto/variations_seed.proto ./proto/study.proto -o proto/generated/proto_bundle.js && pbts -o proto/generated/proto_bundle.d.ts proto/generated/proto_bundle.js",
     "build:tracker": "tsc -p finch_tracker/tsconfig.json",
@@ -14,6 +14,7 @@
     "serve:dev": "webpack serve -c web/webpack.config.js --hot --mode=development"
   },
   "devDependencies": {
+    "@jest/globals": "^29.7.0",
     "@types/node": "^20.5.7",
     "@types/react": "^18.2.21",
     "@typescript-eslint/eslint-plugin": "^6.4.1",
